moved internal logic to lib.rs
